refactor: lotto.domain 패키지로 도메인 및 테스트 클래스 이동
